break out bundler and deploy into different commands
